Cambio de CRUD de Almnos a page
