+++ apps/system/convex/schema.ts
@@ -152,7 +152,7 @@ const applicationTables = {
     apellidos: v.string(),
     fechaNacimiento: v.string(),
     email: v.optional(v.string()),
-    telefono: v.optional(v.number()),
+    telefono: v.optional(v.string()),
     direccion: v.optional(v.string()),
     activo: v.boolean(),
   })
